pyproject.toml
[tool.poetry]
name = "spacedrepetition"
version = "0.1.0"
description = ""
authors = ["oguzhanoxel <oguzhanoksel@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
djangorestframework = "^3.12.2"
psycopg2 = "^2.8.6"
drf-nested-routers = "^0.92.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
